review some code

+++ app/controller/articles.ts
@@ -2,7 +2,6 @@
 import { Controller } from "egg";
 import * as fs from 'fs';
 import * as pump from 'pump';
-import * as formidable from 'formidable';
 import * as difference from 'lodash/difference';
 
 const indexParams = {
@@ -32,36 +31,6 @@ const createParams = {
   content: 'string'
 }
 
-const UpdateInfoParams = {
-  cover: {
-    type: 'string',
-    required: false
-  },
-  category: {
-    type: 'enum',
-    values: ['TECHNICAL', 'LIFE', 'PRIVACY', 'DRAFT'],
-    default: 'DRAFT',
-    required: false
-  },
-  selectTags: {
-    type: 'array',
-    itemType: "string",
-    required: false
-  },
-  newTags: {
-    type: 'array',
-    itemType: "string",
-    required: false
-  }
-}
-
-const CreateTagParams = {
-  name: {
-    type: 'string',
-    reuqired: true,
-    max: 10
-  }
-}
 export default class ArticleController extends Controller{
   // 文章列表
   async index(){
@@ -71,7 +40,7 @@ export default class ArticleController extends Controller{
 
     // 构建查询参数并查询
     const { page, size } = ctx.request.query;
-    const articles = await ctx.service.article.findAll(page, size)
+    const articles = await ctx.service.article.findAll({page, size})
 
     // 返回结果
     ctx.response.success({
@@ -189,55 +158,18 @@ export default class ArticleController extends Controller{
     ctx.response.success()
   }
 
-  // 获取所有标签
-  async tags(){
-    const { ctx } = this;
-
-    const tags = await ctx.service.article.findAllTags();
-    ctx.response.success({
-      data: tags
-    })
-  }
-
-  // 新建标签
-  async createTags(){
-    const { ctx, config } = this;
-    // 验证参数
-    ctx.validate(CreateTagParams, ctx.request.body);
-    const { name } = ctx.request.body;
-
-    const tag = await ctx.service.article.createTag(name);
-    if(!tag){
-      return ctx.response.failure(
-        config.ERR_TYPE.TAG_IS_EXIST
-      )
-    }
-    ctx.response.success({
-      data: tag
-    })
-  }
-
-  parse(req) {
-    const form = new formidable.IncomingForm();
-    return new Promise((resolve, reject) => {
-        form.parse(req, (err, fields, files) => {
-            resolve({ fields, files });
-        });
-    });
-  }
-
   // 更新文章额外信息
   async updateInfo(){
     const { ctx, config } = this;
     const { id } = ctx.params;
     const body = ctx.request.body
 
-    const {fields, files} = await this.parse(ctx.req);
+    // 解析form-data
+    const {fields, files} = await ctx.helper.parse(ctx.req);
     let {category, tags} = fields;
     tags = JSON.parse(fields.tags);
-    // console.log(category, tags)
 
-    // ctx.validate(UpdateInfoParams, body);
+    // 验证 id
     const article = await ctx.service.article.findById(id);
     if(!article){
       return ctx.response.failure(
@@ -247,35 +179,40 @@ export default class ArticleController extends Controller{
 
     // 图片上传
     const file = files.cover;
-    const stream = fs.createReadStream(file._writeStream.path)
-    const dir = await this.service.upload.getUploadFile(file.originalFilename, 'cover');
-    const writeStream = fs.createWriteStream(dir.uploadDir)
-    await pump(stream, writeStream);
+    let dir;
+    if(file){
+      const stream = fs.createReadStream(file._writeStream.path)      
+      dir = await this.service.upload.getUploadFile(file.originalFilename, 'cover');
+      const writeStream = fs.createWriteStream(dir.uploadDir)
+      await pump(stream, writeStream);
+    }
 
     const originCover = article.cover;
     const originTags = article.tags.map(item => String(item._id));
     const addTags = difference(tags, originTags);
     const removeTags = difference(originTags, tags);
-    // console.log(tags, originTags, addTags, removeTags)
 
     await ctx.service.article.updateInfo(id, {
       category,
       tags,
-      cover: dir.saveDir
+      cover: file ? dir.saveDir: ''
     })
+
     // 删除旧图片
-    await fs.unlinkSync(originCover.replace(ctx.origin, 'app'))
+    if(file && originCover){
+      await fs.unlinkSync(originCover.replace(ctx.origin, 'app'))
+    }
     if(addTags.length){
-      await ctx.service.article.tagAddArticle(addTags, id)
+      await ctx.service.tag.tagAddArticle(addTags, id)
     }
     if(removeTags.length){
-      await ctx.service.article.tagRemoveArticle(removeTags, id)
+      await ctx.service.tag.tagRemoveArticle(removeTags, id)
     }
-    const res = await ctx.service.article.findById(id);
+    const data = await ctx.service.article.findById(id);
     // 返回结果
     ctx.response.success({
       message: '更新成功',
-      data: res
+      data: data
     })
   }
 }
